DaySheet report (end of day report) , take 2.
This is a report used by our clients at the end of the day to balance. I
have included 3 versions of the report. The default version is the one
preferred by our clients. I added another selection to the billing
manager that allows this version to be run for a specific user. I have a
pdf file with screen shots and descriptions placed in Documents directory.